+++ test/Servant/API/Routes/RouteSpec.hs
@@ -3,6 +3,7 @@ module Servant.API.Routes.RouteSpec
   )
 where
 
+import Data.Coerce
 import Data.Function
 import qualified Data.Set as Set
 import qualified Data.Text as T
@@ -30,6 +31,8 @@ instance Q.Arbitrary Route where
     _routeRequestBody <- arbitrary
     _routeResponse <- arbitrary
     _routeAuths <- Set.fromList <$> Q.listOf genAuths
+    _routeDescription <- Q.liftArbitrary genDescription
+    _routeSummary <- Q.liftArbitrary genSummary
 
     pure Route {..}
     where
@@ -38,6 +41,8 @@ instance Q.Arbitrary Route where
           [ Basic <$> genAlphaText
           , Custom <$> genAlphaText
           ]
+      genDescription = RouteDescription . T.unwords <$> Q.listOf genAlphaText
+      genSummary = RouteSummary . T.unwords <$> Q.scale (`div` 2) (Q.listOf genAlphaText)
 
   shrink r =
     routeMethod shrinkMethod r
@@ -47,6 +52,8 @@ instance Q.Arbitrary Route where
       <> routeRequestBody Q.shrink r
       <> routeResponse Q.shrink r
       <> routeAuths (shrinkSet shrinkAuth) r
+      <> routeDescription (Q.liftShrink (coerce shrinkText)) r
+      <> routeSummary (Q.liftShrink (coerce shrinkText)) r
     where
       shrinkMethod = either (const []) (fmap renderStdMethod . Q.shrinkBoundedEnum) . parseMethod
       shrinkSet shr = fmap Set.fromList . Q.shrinkList shr . Set.toList