+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Haskell Package Versioning Policy](https://pvp.hask
 
 - A function to sort `Route`s before printing them to stdout. [#32](https://github.com/fpringle/servant-routes/pull/32)
 - More GHC versions in `tested-with`.
+- Add support for Servant's `Description` and `Summary` combinators. [#29](https://github.com/fpringle/servant-routes/pull/29)
 
 ### Fixed